+++ lib/gooddata/cli/commands/auth_cmd.rb
@@ -4,10 +4,8 @@ require_relative '../shared'
 require_relative '../../commands/auth'
 
 GoodData::CLI.module_eval do
-
   desc 'Work with your locally stored credentials'
   command :auth do |c|
-
     c.desc 'Store your credentials to ~/.gooddata so client does not have to ask you every single time'
     c.command :store do |store|
       store.action do |_global_options, _options, _args|
@@ -15,5 +13,4 @@ GoodData::CLI.module_eval do
       end
     end
   end
-
 end